add refresh feature to news list
